Normie added-Armando
I added the normie character in the hero position. Assets and textures added under character folder in a new subfolder for normie. In the subfolder there is also animations for normie. Got the character to move correctly with idle and play through the game. I saved it under a new scene just in case I changed something I wasnt suppose too. But I think it is fine. And made a new branch